Make build compatible with clang on windows

+++ launcher/GZip_test.cpp
@@ -25,7 +25,8 @@ slots:
         QByteArray compressed;
         QByteArray decompressed;
         std::default_random_engine eng((std::random_device())());
-        std::uniform_int_distribution<uint8_t> idis(0, std::numeric_limits<uint8_t>::max());
+        // std::iniform_int_distribution<uint8_t> is not in the standard (fails with MS std lib)
+        std::uniform_int_distribution<uint16_t> idis(0, std::numeric_limits<uint8_t>::max());
 
         // initialize random buffer
         for(int i = 0; i < size; i++)
